Add new Configuration Variable
Add a new Configuration Variable 'REPLACE_REGISTRY_URL'.
This this Variable a User can decide if he want that the the "Target"
Docker Registry Domain (eg. 'quay.io') will be replaced with this own
Domain, or if this Registry Domain will be simply pre-pendeted.

Example:
  Kubelet pulls Image:            'docker.io/library/nginx:latest'
  With 'REPLACE_..' Enabled:      'jfrog.local/library/nginx:latest'
  With 'REPLACE_..' Disabled:     'jfrog.local/docker.io/library/nginx:latest'

This is because we at FABMation are "sorting" the Images from 3rd-Party
Registries (and sign them). So having the possibility to decide if the
original Registry Domain should be replaced or not is a nice Feature.

Signed-off-by: Emanuel Bennici <benniciemanuel78@gmail.com>