Add priorityClass to Helm-Chart
Add priorityClassName to the deployment Template of the Helm-Chart. We
can add this Field without any checks because Kubernetes enables this
(since v1.11) from default settings[1].
And the PriorityClass `system-cluster-critical` is a default
PriorityClass so there would be no problems/ incompatibility.

We also do not need a extra Helm Value/ Switch to let the User decide if
he want `priorityClass'es` because since `kubenab` would be a relevant
Service which the Cluster needs to run correctly, this priorityClass
would be needed by anyone.

Closes #17

[1]: https://kubernetes.io/docs/concepts/configuration/pod-priority-preemption/