Add STRIP_DEBUG Variable to Makefile
Add the STRIP_DEBUG Variable to the Makefile (with Documentation) which
– by default – enables the `strip_debug` Tag to "remove" all the
`log.Debug(.*)` Calls.

The User can decide if he wants that the Debug Messages are striped or
not. For productiv Images it would be better to strip those Debug Log
entries because they slow `kubenab` really down and also blow the Log
Output up.

Signed-off-by: Emanuel Bennici <benniciemanuel78@gmail.com>