Change App Version Info Variable Names in Makefile
Since we have changed the App Info Variables in Commit
d777dd44095968d871631f7057e5026d03b0f7f9 we have to change them also in
the Makefile under the `ldflags` or it would break the Makefile.

This is the last Patch which implements the Enhancement of jfrog#30

Why changing the App Info Variable Names?

Because goreleaser use those Names as from default so we have not to
edit more Fields of the `.goreleaser.yaml` Config.